backend/requirements.txt
alembic==1.7.4
attrs==21.2.0
black==21.9b0
click==8.0.1
config==0.5.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.1
regex==2021.10.8
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
Werkzeug==2.0.1
